+++ Sources/RInAppMessaging/Utilities/AnalyticsTracker.swift
@@ -10,7 +10,11 @@ internal struct AnalyticsTracker {
     static func sendEventName(_ name: Constants.RAnalytics, dataObject: [String: Any] = [:]) {
         var eventData = dataObject
         eventData[Constants.RAnalytics.Keys.deviceID] = UIDevice.deviceID
-        AnalyticsBroadcaster.sendEventName(name.rawValue,
-                                           dataObject: eventData)
+        
+        if let customAccNumber = BundleInfo.rmcRATAccountId, RInAppMessaging.isRMCEnvironment {
+            let rmcEventName = name.rawValue == Constants.RAnalytics.impressionsEventName.rawValue ? Constants.RAnalytics.rmcImpressionsEventName : Constants.RAnalytics.rmcPushPrimerEventName
+            AnalyticsBroadcaster.sendEventName(rmcEventName.rawValue, dataObject: eventData, customAccountNumber: customAccNumber)
+        }
+        AnalyticsBroadcaster.sendEventName(name.rawValue, dataObject: eventData)
     }
 }