+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     - Prevent showing Push Priemr Campaign if Notificaation Permission was granted already [RMCCX-6707]
     - Added CustomJson feature for RMC SDK users [RMCCX-6712] 
     - Redirect user to App Notification Permission Settings if permission was denied previously in Push Primer [RMCCX-6710]
+    - Send RMC impression event to RAT [RMCCX-6939]
+    - Send RMC push primer event to RAT [RMCCX-6938]
 
 ### 8.3.0 (2024-05-13) 
 - Improvements: